aggregation_service/requirements.txt
requests==2.31.0
pymongo==4.6.3
schedule==1.2.1
python-dotenv==1.0.1
